+++ src/StorageBase/Action/IdentityRedirect/Create/IdentityRedirectCreateResultCollection.php
@@ -1,0 +1,32 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\Storage\Base\Action\IdentityRedirect\Create;
+
+use Heptacom\HeptaConnect\Dataset\Base\AttachmentCollection;
+use Heptacom\HeptaConnect\Dataset\Base\Contract\AttachmentAwareInterface;
+use Heptacom\HeptaConnect\Dataset\Base\Support\AbstractObjectCollection;
+use Heptacom\HeptaConnect\Dataset\Base\Support\AttachmentAwareTrait;
+
+/**
+ * @extends AbstractObjectCollection<IdentityRedirectCreateResult>
+ */
+final class IdentityRedirectCreateResultCollection extends AbstractObjectCollection implements AttachmentAwareInterface
+{
+    use AttachmentAwareTrait;
+
+    public function __construct(iterable $items = [])
+    {
+        parent::__construct($items);
+        $this->attachments = new AttachmentCollection();
+    }
+
+    /**
+     * @psalm-return IdentityRedirectCreateResult::class
+     */
+    protected function getT(): string
+    {
+        return IdentityRedirectCreateResult::class;
+    }
+}